pyproject.toml
[project]
name = "l4q"
version = "1.0"
description = "Query Left 4 Dead 2 servers."
dependencies = [
    "Quart>=0.20.0",
    "quart-tasks>=0.5.0",
    "aiodns>=3.5.0",
    "setuptools>=80.4.0",
    "retry>=0.9.2",
    "validators>=0.35.0",
    "python-dotenv>=1.1.0",
    "python-a2s>=1.4.1",
    "l4d2query @ git+https://github.com/Yepoleb/l4d2query@master"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Flask",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Games/Entertainment",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["instance", ".env"]
